Extract out getRxFeed
